Move Flag measurements to utils
