+++ src/mqt/qecc/circuit_synthesis/state_prep.py
@@ -16,10 +16,9 @@ from qiskit.dagcircuit import DAGOutNode
 from ..codes import InvalidCSSCodeError
 from .synthesis_utils import (
     build_css_circuit_from_cnot_list,
-    gaussian_elimination_min_column_ops,
-    gaussian_elimination_min_parallel_eliminations,
     heuristic_gaussian_elimination,
     iterative_search_with_timeout,
+    measure_flagged,
     odd_overlap,
     optimal_elimination,
     run_with_timeout,
@@ -34,7 +33,7 @@ if TYPE_CHECKING:  # pragma: no cover
     from collections.abc import Callable
 
     import numpy.typing as npt
-    from qiskit import AncillaQubit, ClBit, DagCircuit, DAGNode, Qubit
+    from qiskit import DagCircuit, DAGNode
     from qiskit.quantum_info import PauliList
 
     from ..codes import CSSCode
@@ -237,86 +236,6 @@ def heuristic_prep_circuit(code: CSSCode, optimize_depth: bool = True, zero_stat
     checks, cnots = heuristic_gaussian_elimination(checks, parallel_elimination=optimize_depth)
 
     circ = _build_state_prep_circuit_from_back(checks, cnots, zero_state)
-    return StatePrepCircuit(circ, code, zero_state)
-
-
-def _generate_circ_with_bounded_depth(
-    checks: npt.NDArray[np.int8], max_depth: int, zero_state: bool = True
-) -> QuantumCircuit | None:
-    assert max_depth > 0, "max_depth should be greater than 0"
-    res = gaussian_elimination_min_parallel_eliminations(checks, _final_matrix_constraint, max_depth)
-    if res is None:
-        return None
-    checks, cnots = res
-    return _build_state_prep_circuit_from_back(checks, cnots, zero_state)
-
-
-def _generate_circ_with_bounded_gates(
-    checks: npt.NDArray[np.int8], max_depth: int, zero_state: bool = True
-) -> QuantumCircuit | None:
-    assert max_depth > 0, "max_depth should be greater than 0"
-    res = gaussian_elimination_min_column_ops(checks, _final_matrix_constraint, max_depth)
-    if res is None:
-        return None
-    checks, cnots = res
-    return _build_state_prep_circuit_from_back(checks, cnots, zero_state)
-
-
-def _optimal_circuit(
-    code: CSSCode,
-    prep_func: Callable[[npt.NDArray[np.int8], int, bool], QuantumCircuit | None],
-    zero_state: bool = True,
-    min_param: int = 1,
-    max_param: int = 10,
-    min_timeout: int = 1,
-    max_timeout: int = 3600,
-) -> StatePrepCircuit | None:
-    """Synthesize a state preparation circuit for a CSS code that minimizes the circuit w.r.t. some metric param according to prep_func.
-
-    Args:
-        code: The CSS code to prepare the state for.
-        zero_state: If True, prepare the +1 eigenstate of the Z basis. If False, prepare the +1 eigenstate of the X basis.
-        prep_func: The function to optimize the circuit with respect to.
-        min_param: minimum parameter to start with
-        max_param: maximum parameter to reach
-        min_timeout: minimum timeout to start with
-        max_timeout: maximum timeout to reach
-    """
-    if code.Hx is None or code.Hz is None:
-        msg = "Code must have both X and Z stabilizers defined."
-        raise ValueError(msg)
-    checks = code.Hx if zero_state else code.Hz
-
-    def fun(param: int) -> QuantumCircuit | None:
-        return prep_func(checks, param, zero_state)
-
-    res = iterative_search_with_timeout(
-        fun,
-        min_param,
-        max_param,
-        min_timeout,
-        max_timeout,
-    )
-
-    if res is None:
-        return None
-    circ, curr_param = res
-    if circ is None:
-        return None
-
-    logging.info(f"Solution found with param {curr_param}")
-    # Solving a SAT instance is much faster than proving unsat in this case
-    # so we iterate backwards until we find an unsat instance or hit a timeout
-    logging.info("Trying to minimize param")
-    while True:
-        logging.info(f"Trying param {curr_param - 1}")
-        opt_res = run_with_timeout(fun, curr_param - 1, timeout=max_timeout)
-        if opt_res is None or (isinstance(opt_res, str) and opt_res == "timeout"):
-            break
-        circ = opt_res
-        curr_param -= 1
-
-    logging.info(f"Optimal param: {curr_param}")
     return StatePrepCircuit(circ, code, zero_state)
 
 
@@ -971,348 +890,6 @@ def naive_verification_circuit(sp_circ: StatePrepCircuit) -> QuantumCircuit:
     x_measurements = list(sp_circ.code.Hz)
     reps = sp_circ.max_errors
     return _measure_ft_stabs(sp_circ, z_measurements * reps, x_measurements * reps)
-
-
-def w_flag_pattern(w: int) -> list[int]:
-    """Return the w-flag construction from https://arxiv.org/abs/1708.02246.
-
-    Args:
-        w: The number of w-flags to construct.
-
-    Returns:
-        The w-flag pattern.
-    """
-    s1 = [2 * j + 2 for j in reversed(range((w - 4) // 2))]
-    s2 = [w - 3, 0]
-    s3 = [2 * j + 1 for j in reversed(range((w - 4) // 2))]
-    return s1 + s2 + s3 + [w - 2]
-
-
-def _ancilla_cnot(qc: QuantumCircuit, qubit: Qubit | AncillaQubit, ancilla: AncillaQubit, z_measurement: bool) -> None:
-    if z_measurement:
-        qc.cx(qubit, ancilla)
-    else:
-        qc.cx(ancilla, qubit)
-
-
-def _flag_measure(qc: QuantumCircuit, flag: AncillaQubit, meas_bit: ClBit, z_measurement: bool) -> None:
-    if z_measurement:
-        qc.h(flag)
-    qc.measure(flag, meas_bit)
-
-
-def _flag_reset(qc: QuantumCircuit, flag: AncillaQubit, z_measurement: bool) -> None:
-    qc.reset(flag)
-    if z_measurement:
-        qc.h(flag)
-
-
-def _flag_init(qc: QuantumCircuit, flag: AncillaQubit, z_measurement: bool) -> None:
-    if z_measurement:
-        qc.h(flag)
-
-
-def _measure_stab_unflagged(
-    qc: QuantumCircuit,
-    stab: list[Qubit] | npt.NDArray[np.int_],
-    ancilla: AncillaQubit,
-    measurement_bit: ClBit,
-    z_measurement: bool = True,
-) -> None:
-    if not z_measurement:
-        qc.h(ancilla)
-        qc.cx([ancilla] * len(stab), stab)
-        qc.h(ancilla)
-    else:
-        qc.cx(stab, [ancilla] * len(stab))
-    qc.measure(ancilla, measurement_bit)
-
-
-def measure_flagged(
-    qc: QuantumCircuit,
-    stab: list[Qubit] | npt.NDArray[np.int_],
-    ancilla: AncillaQubit,
-    measurement_bit: ClBit,
-    t: int,
-    z_measurement: bool = True,
-) -> None:
-    """Measure a w-flagged stabilizer with the general scheme.
-
-    The measurement is done in place.
-
-    Args:
-        qc: The quantum circuit to add the measurement to.
-        stab: The qubits to measure.
-        ancilla: The ancilla qubit to use for the measurement.
-        measurement_bit: The classical bit to store the measurement result of the ancilla.
-        t: The number of errors to protect from.
-        z_measurement: Whether to measure an X (False) or Z (True) stabilizer.
-    """
-    w = len(stab)
-    if t == 1:
-        _measure_stab_one_flagged(qc, stab, ancilla, measurement_bit, z_measurement)
-        return
-
-    if w < 3 and t == 2:
-        _measure_stab_unflagged(qc, stab, ancilla, measurement_bit, z_measurement)
-        return
-
-    if w == 4 and t == 2:
-        measure_flagged_4(qc, stab, ancilla, measurement_bit, z_measurement)
-        return
-
-    if w == 6 and t == 2:
-        measure_flagged_6(qc, stab, ancilla, measurement_bit, z_measurement)
-        return
-
-    if w == 8 and t == 2:
-        measure_flagged_8(qc, stab, ancilla, measurement_bit, z_measurement)
-        return
-
-    if t == 2:
-        measure_stab_two_flagged(qc, stab, ancilla, measurement_bit, z_measurement)
-        return
-
-    msg = f"Flagged measurement for w={w} and t={t} not implemented."
-    raise NotImplementedError(msg)
-
-
-def _measure_stab_one_flagged(
-    qc: QuantumCircuit,
-    stab: list[Qubit] | npt.NDArray[np.int_],
-    ancilla: AncillaQubit,
-    measurement_bit: ClBit,
-    z_measurement: bool = True,
-) -> None:
-    """Measure a 1-flagged stabilizer using an optimized scheme."""
-    flag_reg = AncillaRegister(1)
-    meas_reg = ClassicalRegister(1)
-    qc.add_register(flag_reg)
-    qc.add_register(meas_reg)
-    flag = flag_reg[0]
-    flag_meas = meas_reg[0]
-    if not z_measurement:
-        qc.h(ancilla)
-
-    _ancilla_cnot(qc, stab[0], ancilla, z_measurement)
-    _flag_init(qc, flag, z_measurement)
-
-    _ancilla_cnot(qc, flag, ancilla, z_measurement)
-
-    for q in stab[1:-1]:
-        _ancilla_cnot(qc, q, ancilla, z_measurement)
-
-    _ancilla_cnot(qc, flag, ancilla, z_measurement)
-    _flag_measure(qc, flag, flag_meas, z_measurement)
-
-    _ancilla_cnot(qc, stab[-1], ancilla, z_measurement)
-
-    if not z_measurement:
-        qc.h(ancilla)
-    qc.measure(ancilla, measurement_bit)
-
-
-def measure_stab_two_flagged(
-    qc: QuantumCircuit,
-    stab: list[Qubit] | npt.NDArray[np.int_],
-    ancilla: AncillaQubit,
-    measurement_bit: ClBit,
-    z_measurement: bool = True,
-) -> None:
-    """Measure a 2-flagged stabilizer using the scheme of https://arxiv.org/abs/1708.02246 (page 13)."""
-    assert len(stab) > 4
-    n_flags = (len(stab) + 1) // 2 - 1
-    flag_reg = AncillaRegister(n_flags)
-    meas_reg = ClassicalRegister(n_flags)
-
-    qc.add_register(flag_reg)
-    qc.add_register(meas_reg)
-
-    if not z_measurement:
-        qc.h(ancilla)
-
-    _ancilla_cnot(qc, stab[0], ancilla, z_measurement)
-
-    _flag_init(qc, flag_reg[0], z_measurement)
-    _ancilla_cnot(qc, flag_reg[0], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, stab[1], ancilla, z_measurement)
-    _flag_init(qc, flag_reg[1], z_measurement)
-    _ancilla_cnot(qc, flag_reg[1], ancilla, z_measurement)
-
-    cnots = 2
-    flags = 2
-    for q in stab[2:-2]:
-        _ancilla_cnot(qc, q, ancilla, z_measurement)
-        cnots += 1
-        if cnots % 2 == 0 and cnots < len(stab) - 2:
-            _flag_init(qc, flag_reg[flags], z_measurement)
-            _ancilla_cnot(qc, flag_reg[flags], ancilla, z_measurement)
-        if cnots >= 7 and cnots % 2 == 1:
-            _ancilla_cnot(qc, flag_reg[flags - 1], ancilla, z_measurement)
-            _flag_measure(qc, flag_reg[flags - 1], meas_reg[flags - 1], z_measurement)
-        flags += 1
-
-    _ancilla_cnot(qc, flag_reg[0], ancilla, z_measurement)
-    _flag_measure(qc, flag_reg[0], meas_reg[0], z_measurement)
-
-    _ancilla_cnot(qc, stab[-2], ancilla, z_measurement)
-
-    cnots += 1
-    if cnots >= 7 and cnots % 2 == 1:
-        _ancilla_cnot(qc, flag_reg[flags - 1], ancilla, z_measurement)
-        _flag_measure(qc, flag_reg[flags - 1], meas_reg[flags - 1], z_measurement)
-
-    _ancilla_cnot(qc, flag_reg[1], ancilla, z_measurement)
-    _flag_measure(qc, flag_reg[1], meas_reg[1], z_measurement)
-
-    _ancilla_cnot(qc, stab[-1], ancilla, z_measurement)
-
-    cnots += 1
-    if cnots >= 7 and cnots % 2 == 1:
-        _ancilla_cnot(qc, flag_reg[flags - 1], ancilla, z_measurement)
-        _flag_measure(qc, flag_reg[flags - 1], meas_reg[flags - 1], z_measurement)
-    if not z_measurement:
-        qc.h(ancilla)
-
-    qc.measure(ancilla, measurement_bit)
-
-
-def measure_flagged_4(
-    qc: QuantumCircuit,
-    stab: list[Qubit] | npt.NDArray[np.int_],
-    ancilla: AncillaQubit,
-    measurement_bit: ClBit,
-    z_measurement: bool = True,
-) -> None:
-    """Measure a 4-flagged stabilizer using an optimized scheme."""
-    assert len(stab) == 4
-    flag_reg = AncillaRegister(1)
-    meas_reg = ClassicalRegister(1)
-    qc.add_register(flag_reg)
-    qc.add_register(meas_reg)
-    flag = flag_reg[0]
-    flag_meas = meas_reg[0]
-
-    if not z_measurement:
-        qc.h(ancilla)
-
-    _ancilla_cnot(qc, stab[0], ancilla, z_measurement)
-    _flag_init(qc, flag, z_measurement)
-
-    _ancilla_cnot(qc, flag, ancilla, z_measurement)
-
-    _ancilla_cnot(qc, stab[1], ancilla, z_measurement)
-    _ancilla_cnot(qc, stab[2], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, flag, ancilla, z_measurement)
-    _flag_measure(qc, flag, flag_meas, z_measurement)
-
-    _ancilla_cnot(qc, stab[3], ancilla, z_measurement)
-
-    if not z_measurement:
-        qc.h(ancilla)
-    qc.measure(ancilla, measurement_bit)
-
-
-def measure_flagged_6(
-    qc: QuantumCircuit,
-    stab: list[Qubit] | npt.NDArray[np.int_],
-    ancilla: AncillaQubit,
-    measurement_bit: ClBit,
-    z_measurement: bool = True,
-) -> None:
-    """Measure a 6-flagged stabilizer using an optimized scheme."""
-    assert len(stab) == 6
-    flag = AncillaRegister(2)
-    meas = ClassicalRegister(2)
-
-    qc.add_register(flag)
-    qc.add_register(meas)
-
-    if not z_measurement:
-        qc.h(ancilla)
-
-    _ancilla_cnot(qc, stab[0], ancilla, z_measurement)
-
-    _flag_init(qc, flag[0], z_measurement)
-    _ancilla_cnot(qc, flag[0], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, stab[1], ancilla, z_measurement)
-
-    _flag_init(qc, flag[1], z_measurement)
-    _ancilla_cnot(qc, flag[1], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, stab[2], ancilla, z_measurement)
-    _ancilla_cnot(qc, stab[3], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, flag[0], ancilla, z_measurement)
-    _flag_measure(qc, flag[0], meas[0], z_measurement)
-
-    _ancilla_cnot(qc, stab[4], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, flag[1], ancilla, z_measurement)
-    _flag_measure(qc, flag[1], meas[1], z_measurement)
-
-    _ancilla_cnot(qc, stab[5], ancilla, z_measurement)
-
-    if not z_measurement:
-        qc.h(ancilla)
-    qc.measure(ancilla, measurement_bit)
-
-
-def measure_flagged_8(
-    qc: QuantumCircuit,
-    stab: list[Qubit] | npt.NDArray[np.int_],
-    ancilla: AncillaQubit,
-    measurement_bit: ClBit,
-    z_measurement: bool = True,
-) -> None:
-    """Measure an 8-flagged stabilizer using an optimized scheme."""
-    assert len(stab) == 8
-    flag = AncillaRegister(3)
-    meas = ClassicalRegister(3)
-    qc.add_register(flag)
-    qc.add_register(meas)
-
-    if not z_measurement:
-        qc.h(ancilla)
-
-    _ancilla_cnot(qc, stab[0], ancilla, z_measurement)
-
-    _flag_init(qc, flag[0], z_measurement)
-    _ancilla_cnot(qc, flag[0], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, stab[1], ancilla, z_measurement)
-
-    _flag_init(qc, flag[1], z_measurement)
-    _ancilla_cnot(qc, flag[1], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, stab[2], ancilla, z_measurement)
-    _ancilla_cnot(qc, stab[3], ancilla, z_measurement)
-
-    _flag_init(qc, flag[2], z_measurement)
-    _ancilla_cnot(qc, flag[2], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, stab[4], ancilla, z_measurement)
-    _ancilla_cnot(qc, stab[5], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, flag[0], ancilla, z_measurement)
-    _flag_measure(qc, flag[0], meas[0], z_measurement)
-
-    _ancilla_cnot(qc, stab[6], ancilla, z_measurement)
-
-    _ancilla_cnot(qc, flag[2], ancilla, z_measurement)
-    _flag_measure(qc, flag[2], meas[2], z_measurement)
-
-    _ancilla_cnot(qc, flag[1], ancilla, z_measurement)
-    _flag_measure(qc, flag[1], meas[1], z_measurement)
-
-    _ancilla_cnot(qc, stab[7], ancilla, z_measurement)
-
-    if not z_measurement:
-        qc.h(ancilla)
-    qc.measure(ancilla, measurement_bit)
 
 
 def _hook_errors(stabs: list[npt.NDArray[np.int8]]) -> npt.NDArray[np.int8]:
